feat: 그래프 기능 구현

+++ app/src/main/java/com/example/a2024capstonesample/data/PhotoData.kt
@@ -3,5 +3,5 @@ package com.example.a2024capstonesample.data
 data class PhotoData(
     val date: String, // 사진 촬영 날짜
     val photoPath: String, // 사진 경로
-    var measurement: Float // 분석 수치
+    var measurement: Double // 분석 수치
 )
